Refractoring the UI for  mapping service for nursing
